apply formatter with `forge fmt`

+++ src/FBA.sol
@@ -22,14 +22,16 @@ contract FBA {
     Fill[] public fills;
 
     struct Fill {
-        uint amount;
-        uint price;
+        uint256 amount;
+        uint256 price;
     }
+
     struct PlaceResult {
         uint256 price;
         bool side;
         uint256 amount;
     }
+
     struct CancelResult {
         uint256 price;
         bool side;
@@ -48,57 +50,24 @@ contract FBA {
 
     function initFBA() external returns (bytes memory) {
         // For the array
-        Suave.DataRecord memory bidArr = Suave.newDataRecord(
-            0,
-            addressList,
-            addressList,
-            "suaveFBA:v0:dataId"
-        );
-        FBAHeap.ArrayMetadata memory bidAm = FBAHeap.ArrayMetadata(
-            0,
-            bidArr.id
-        );
+        Suave.DataRecord memory bidArr = Suave.newDataRecord(0, addressList, addressList, "suaveFBA:v0:dataId");
+        FBAHeap.ArrayMetadata memory bidAm = FBAHeap.ArrayMetadata(0, bidArr.id);
         FBAHeap.arrSetMetadata(bidAm);
 
-        Suave.DataRecord memory askArr = Suave.newDataRecord(
-            0,
-            addressList,
-            addressList,
-            "suaveFBA:v0:dataId"
-        );
-        FBAHeap.ArrayMetadata memory askAm = FBAHeap.ArrayMetadata(
-            0,
-            askArr.id
-        );
+        Suave.DataRecord memory askArr = Suave.newDataRecord(0, addressList, addressList, "suaveFBA:v0:dataId");
+        FBAHeap.ArrayMetadata memory askAm = FBAHeap.ArrayMetadata(0, askArr.id);
         FBAHeap.arrSetMetadata(askAm);
 
         // For the map
-        Suave.DataRecord memory bidMap = Suave.newDataRecord(
-            0,
-            addressList,
-            addressList,
-            "suaveFBA:v0:dataId"
-        );
+        Suave.DataRecord memory bidMap = Suave.newDataRecord(0, addressList, addressList, "suaveFBA:v0:dataId");
         FBAHeap.MapMetadata memory bidMm = FBAHeap.MapMetadata(bidMap.id);
         FBAHeap.mapSetMetadata(bidMm);
 
-        Suave.DataRecord memory askMap = Suave.newDataRecord(
-            0,
-            addressList,
-            addressList,
-            "suaveFBA:v0:dataId"
-        );
+        Suave.DataRecord memory askMap = Suave.newDataRecord(0, addressList, addressList, "suaveFBA:v0:dataId");
         FBAHeap.MapMetadata memory askMm = FBAHeap.MapMetadata(askMap.id);
         FBAHeap.mapSetMetadata(askMm);
 
-        return
-            abi.encodeWithSelector(
-                this.initFBACallback.selector,
-                bidArr.id,
-                askArr.id,
-                bidMap.id,
-                askMap.id
-            );
+        return abi.encodeWithSelector(this.initFBACallback.selector, bidArr.id, askArr.id, bidMap.id, askMap.id);
     }
 
     function displayFills(Fill[] memory _fills) public payable {
@@ -111,16 +80,10 @@ contract FBA {
     /**
      * @notice Allows user to place a new order and immediately checks for fills
      */
-    function placeOrder(
-        FBAHeap.FBAOrder memory ord
-    ) external returns (bytes memory) {
-        FBAHeap.ArrayMetadata memory bidAm = FBAHeap.arrGetMetadata(
-            bidArrayRef
-        );
+    function placeOrder(FBAHeap.FBAOrder memory ord) external returns (bytes memory) {
+        FBAHeap.ArrayMetadata memory bidAm = FBAHeap.arrGetMetadata(bidArrayRef);
         FBAHeap.MapMetadata memory bidMm = FBAHeap.mapGetMetadata(bidMapRef);
-        FBAHeap.ArrayMetadata memory askAm = FBAHeap.arrGetMetadata(
-            askArrayRef
-        );
+        FBAHeap.ArrayMetadata memory askAm = FBAHeap.arrGetMetadata(askArrayRef);
         FBAHeap.MapMetadata memory askMm = FBAHeap.mapGetMetadata(askMapRef);
 
         // Add it to bids or asks heap...
@@ -132,77 +95,49 @@ contract FBA {
 
         // Assuming order placement was always successful?
         PlaceResult memory pr = PlaceResult(ord.price, ord.side, ord.amount);
-        return
-            abi.encodeWithSelector(this.placeOrderCallback.selector, pr);
+        return abi.encodeWithSelector(this.placeOrderCallback.selector, pr);
     }
 
     /**
      * @notice Allows user to cancel an order they previously placed
      */
-    function cancelOrder(
-        string memory clientId,
-        bool side
-    ) external returns (bytes memory) {
+    function cancelOrder(string memory clientId, bool side) external returns (bytes memory) {
         FBAHeap.FBAOrder memory ord;
 
         if (side == ISBUY) {
             bool maxHeapBids = true;
-            FBAHeap.ArrayMetadata memory bidAm = FBAHeap.arrGetMetadata(
-                bidArrayRef
-            );
-            FBAHeap.MapMetadata memory bidMm = FBAHeap.mapGetMetadata(
-                bidMapRef
-            );
+            FBAHeap.ArrayMetadata memory bidAm = FBAHeap.arrGetMetadata(bidArrayRef);
+            FBAHeap.MapMetadata memory bidMm = FBAHeap.mapGetMetadata(bidMapRef);
             ord = FBAHeap.deleteOrder(maxHeapBids, bidAm, bidMm, clientId);
         } else if (side == ISSELL) {
             bool maxHeapAsks = false;
-            FBAHeap.ArrayMetadata memory askAm = FBAHeap.arrGetMetadata(
-                askArrayRef
-            );
-            FBAHeap.MapMetadata memory askMm = FBAHeap.mapGetMetadata(
-                askMapRef
-            );
+            FBAHeap.ArrayMetadata memory askAm = FBAHeap.arrGetMetadata(askArrayRef);
+            FBAHeap.MapMetadata memory askMm = FBAHeap.mapGetMetadata(askMapRef);
             ord = FBAHeap.deleteOrder(maxHeapAsks, askAm, askMm, clientId);
         }
 
-        CancelResult memory orderResult = CancelResult(
-            ord.price,
-            ord.side,
-            ord.amount
-        );
-        return
-            abi.encodeWithSelector(
-                this.cancelOrderCallback.selector,
-                orderResult
-            );
+        CancelResult memory orderResult = CancelResult(ord.price, ord.side, ord.amount);
+        return abi.encodeWithSelector(this.cancelOrderCallback.selector, orderResult);
     }
 
     function executeFills() external returns (bytes memory) {
-        FBAHeap.ArrayMetadata memory bidAm = FBAHeap.arrGetMetadata(
-            bidArrayRef
-        );
+        FBAHeap.ArrayMetadata memory bidAm = FBAHeap.arrGetMetadata(bidArrayRef);
         FBAHeap.MapMetadata memory bidMm = FBAHeap.mapGetMetadata(bidMapRef);
-        FBAHeap.ArrayMetadata memory askAm = FBAHeap.arrGetMetadata(
-            askArrayRef
-        );
+        FBAHeap.ArrayMetadata memory askAm = FBAHeap.arrGetMetadata(askArrayRef);
         FBAHeap.MapMetadata memory askMm = FBAHeap.mapGetMetadata(askMapRef);
 
-        uint bidFallbackPrice = 0;
-        uint askFallbackPrice = type(uint).max;
-        FBAHeap.FBAOrder memory bestBid = FBAHeap.peekTopOne(
-            bidAm,
-            bidFallbackPrice,
-            ISBUY
-        );
+        uint256 bidFallbackPrice = 0;
+        uint256 askFallbackPrice = type(uint256).max;
+        FBAHeap.FBAOrder memory bestBid = FBAHeap.peekTopOne(bidAm, bidFallbackPrice, ISBUY);
         // asks and bids are the orders that will be possibly matched
         FBAHeap.FBAOrder[] memory asks = FBAHeap.peekTopList(askAm, bestBid.price, ISSELL, askFallbackPrice);
         FBAHeap.FBAOrder[] memory bids = FBAHeap.peekTopList(bidAm, asks[0].price, ISBUY, bidFallbackPrice);
 
         // First part: match orders with the same price
-        for (uint i = 0; i < bids.length; i++) {
-            uint fillAmount;
-            uint fillPrice;
-            for (uint j = 0; j < asks.length; j++) {
+        for (uint256 i = 0; i < bids.length; i++) {
+            uint256 fillAmount;
+            uint256 fillPrice;
+            for (uint256 j = 0; j < asks.length; j++) {
                 if (bids[i].price == asks[j].price) {
                     fillPrice = bids[i].price;
                     if (bids[i].amount > asks[j].amount) {
@@ -248,9 +183,7 @@ contract FBA {
         return abi.encodeWithSelector(this.executeFillsCallback.selector, fills);
     }
 
-    function executeFillsCallback(
-        Fill[] memory _fills
-    ) public payable {
+    function executeFillsCallback(Fill[] memory _fills) public payable {
         displayFills(_fills);
     }
 
@@ -268,23 +201,11 @@ contract FBA {
         bidMapRef = _bidMapRef;
     }
 
-    function placeOrderCallback(
-        PlaceResult memory orderResult
-    ) public payable {
-        emit OrderPlace(
-            orderResult.price,
-            orderResult.side,
-            orderResult.amount
-        );
+    function placeOrderCallback(PlaceResult memory orderResult) public payable {
+        emit OrderPlace(orderResult.price, orderResult.side, orderResult.amount);
     }
 
-    function cancelOrderCallback(
-        CancelResult memory orderResult
-    ) public payable {
-        emit OrderCancel(
-            orderResult.price,
-            orderResult.side,
-            orderResult.amount
-        );
+    function cancelOrderCallback(CancelResult memory orderResult) public payable {
+        emit OrderCancel(orderResult.price, orderResult.side, orderResult.amount);
     }
 }
